[5249] Updated control plane for delay server migration.
The control plane now takes an external atomic boolean on construction
which allows it to report whether it is accepting connections. This is
required so that the leader/successor algorithm can determine if a delay
server is running on the leader (i.e. the remote host).

The control plane also uses irods::get_delay_server_pid() instead of
irods::server_properties and irods::RE_PID_KW. This change is required
because the delay server relies on a PID file.